+++ example/lib/json_data.dart
@@ -2,6 +2,15 @@ List<int> listJsonData() {
   return List.generate(200, (index) => index);
 }
 
+Map<String, dynamic> largeJsonData() {
+  return {
+    'items': List.generate(
+      230 * 100 * 100,
+      (index) => {'index': index},
+    ),
+  };
+}
+
 Map<String, dynamic> getJsonData() {
   return {
     'root': {
@@ -32,10 +41,6 @@ Map<String, dynamic> getJsonData() {
       },
       'largeText':
           '  sedi tempero lemiscus poena tenus corrumpo Inda incredibilis centum putus dulcedo medius quarum aliquotiens opprobrium promptus carbo tendo. minimus non ius instructus uberrime extorqueo praetorgredior consto do cotidie compleo protraho exercitus poena frango Brocherota subvenio fidens Werumensium decet. presumo indigeo forsit directus infortunium neque vulgaris conturbo plagiarius socius Berlinmonte innotesco facina praemo itaque sequi.',
-      'largeList': List.generate(
-        230 * 100 * 100,
-        (index) => {'index': index},
-      ),
     },
   };
 }